+++ wcsaxes/tests/test_transform_coord_meta.py
@@ -14,6 +14,8 @@ from .. import WCSAxes
 from .test_images import BaseImageTests
 from ..transforms import CurvedTransform
 
+from . import baseline_dir
+
 requires_astropy_10 = pytest.mark.skipif(str(LooseVersion(astropy.__version__) < LooseVersion("1.0")))
 
 # Create fake transforms that roughly mimic a polar projection
@@ -58,7 +60,7 @@ class LonLatToDistance(CurvedTransform):
 
 class TestTransformCoordMeta(BaseImageTests):
 
-    @pytest.mark.mpl_image_compare(filename='coords_overlay.png', tolerance=1.5)
+    @pytest.mark.mpl_image_compare(baseline_dir=baseline_dir, filename='coords_overlay.png', tolerance=1.5)
     def test_coords_overlay(self):
 
         # Set up a simple WCS that maps pixels to non-projected distances
@@ -106,7 +108,7 @@ class TestTransformCoordMeta(BaseImageTests):
         return fig
 
     @requires_astropy_10
-    @pytest.mark.mpl_image_compare(filename='coords_overlay_auto_coord_meta.png', tolerance=1.5)
+    @pytest.mark.mpl_image_compare(baseline_dir=baseline_dir, filename='coords_overlay_auto_coord_meta.png', tolerance=1.5)
     def test_coords_overlay_auto_coord_meta(self):
 
         fig = plt.figure(figsize=(4, 4))
@@ -128,7 +130,7 @@ class TestTransformCoordMeta(BaseImageTests):
 
         return fig
 
-    @pytest.mark.mpl_image_compare(filename='direct_init.png', tolerance=1.5)
+    @pytest.mark.mpl_image_compare(baseline_dir=baseline_dir, filename='direct_init.png', tolerance=1.5)
     def test_direct_init(self):
 
         s = DistanceToLonLat(R=6378.273)