+++ CHANGES.md
@@ -1,4 +1,4 @@
-0.7 (unreleased)
+0.7 (2016-05-06)
 ----------------
 
 - Removed support for python 2.6.